requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile pyproject.toml
#
cardiac-geometries==0.11.1
    # via ciucci-2024 (pyproject.toml)
click==8.1.7
    # via rich-click
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
daiquiri==3.2.5.1
    # via fenics-pulse
fenics-pulse==2023.2.0
    # via ciucci-2024 (pyproject.toml)
fonttools==4.48.1
    # via matplotlib
h5py==3.9.0
    # via
    #   ciucci-2024 (pyproject.toml)
    #   fenics-pulse
    #   ldrb
kiwisolver==1.4.5
    # via matplotlib
ldrb==2023.5.0
    # via ciucci-2024 (pyproject.toml)
llvmlite==0.40.1
    # via numba
markdown-it-py==3.0.0
    # via rich
matplotlib==3.8.2
    # via
    #   ciucci-2024 (pyproject.toml)
    #   seaborn
mdurl==0.1.2
    # via markdown-it-py
meshio==5.3.5
    # via cardiac-geometries
numba==0.57.1
    # via ldrb
numpy==1.24.4
    # via
    #   cardiac-geometries
    #   contourpy
    #   fenics-pulse
    #   h5py
    #   ldrb
    #   matplotlib
    #   meshio
    #   numba
    #   pandas
    #   scipy
    #   seaborn
packaging==23.2
    # via matplotlib
pandas==2.2.0
    # via
    #   ciucci-2024 (pyproject.toml)
    #   seaborn
pillow==10.2.0
    # via matplotlib
polars==0.20.7
    # via ciucci-2024 (pyproject.toml)
pygments==2.17.2
    # via rich
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-json-logger==2.0.7
    # via daiquiri
pytz==2024.1
    # via pandas
rich==13.7.0
    # via
    #   meshio
    #   rich-click
rich-click==1.7.3
    # via cardiac-geometries
scipy==1.12.0
    # via fenics-pulse
seaborn==0.13.2
    # via ciucci-2024 (pyproject.toml)
six==1.16.0
    # via python-dateutil
typing-extensions==4.9.0
    # via rich-click
tzdata==2023.4
    # via pandas
